Issues 60,62,64,65,67

+++ backend/app/api/services/section_services.py
@@ -1,15 +1,19 @@
-from backend.app.api.utils.responses import NoContent, NotFound
-from backend.app.api.utils.utilities import check_if_guest, check_if_student_or_guest,format_section_details
+from backend.app.api.utils.responses import NoContent, NotFound, Unauthorized
+from backend.app.api.utils.utilities import check_if_guest, check_if_student_or_guest, format_section_details, \
+    check_for_creator
 from backend.app import data
 
 
 async def new_section(user, section_data):
     check_if_student_or_guest(user)
 
-    data.database.insert_query('INSERT INTO sections(title, content, description,information, course_id) VALUES(%s,%s,%s,%s,%s)',
-                 (section_data.title, section_data.content,
-                  section_data.description, section_data.information, section_data.course_id))
+    if check_for_creator(user_id=user.get('id'), course_id=section_data.course_id) == False:
+        raise Unauthorized
 
+    data.database.insert_query(
+        'INSERT INTO sections(title, content, description,information, course_id) VALUES(%s,%s,%s,%s,%s)',
+        (section_data.title, section_data.content,
+         section_data.description, section_data.information, section_data.course_id))
 
     return 'New Section created'
 
@@ -28,8 +32,8 @@ async def sections(user, course_id):
 async def section(user, section_id, course_id):
     check_if_guest(user)
 
-
-    info = data.database.read_query('SELECT * FROM sections WHERE section_id = %s AND course_id = %s', (section_id,course_id))
+    info = data.database.read_query('SELECT * FROM sections WHERE section_id = %s AND course_id = %s',
+                                    (section_id, course_id))
     if info:
         return format_section_details(info)
 
@@ -37,9 +41,11 @@ async def section(user, section_id, course_id):
         raise NoContent
 
 
-async def remove_section(user, section_id):
+async def remove_section(user,course_id, section_id):
     check_if_student_or_guest(user)
 
+    if check_for_creator(user_id=user.get('id'), course_id=course_id) == False:
+        raise Unauthorized
 
     if data.database.read_query('SELECT * FROM sections WHERE section_id = %s', (section_id,)):
         data.database.update_query('DELETE FROM sections WHERE section_id = %s', (section_id,))
@@ -47,4 +53,3 @@ async def remove_section(user, section_id):
 
     else:
         raise NotFound
-
